pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "msk-warp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "warp-lang>=1.5",
    "mujoco>=3.2",
    "tensorboardX",
    "pyyaml",
    "numpy",
]

[tool.setuptools.packages.find]
include = ["msk_warp*"]

[project.optional-dependencies]
dev = ["pytest"]
